refactor(emitter): split ES5 spread array helpers
